Move hyperlink getter code into ia2utils so it can be reused.
